Sim makes Checker[], Checker.getLine(), CheckerEven
Sim.java's main method makes an array of Checkers using a set number of
normal lanes (int normLanes) and express lanes (int expressLanes). It
then makes a single ShopperEvent, CheckerEvents for each Checker in the
array, and starts the domino that is the agenda. NOTE: has no ending
criteria yet.